Move VariableBlockLightHandler methods from WrappedWorldMixin to WorldMixin Add default, null-returning implementations of getChunkAtImmediately and getAnyChunkImmediately to ExtendedWorld
