+++ src/main/java/ca/spottedleaf/starlight/common/world/ExtendedWorld.java
@@ -7,10 +7,14 @@ import net.minecraft.world.chunk.IChunk;
 public interface ExtendedWorld {
 
     // rets full chunk without blocking
-    public Chunk getChunkAtImmediately(final int chunkX, final int chunkZ);
+    default public Chunk getChunkAtImmediately(final int chunkX, final int chunkZ) {
+        return null;
+    }
 
     // rets chunk at any stage, if it exists, immediately
-    public IChunk getAnyChunkImmediately(final int chunkX, final int chunkZ);
+    default public IChunk getAnyChunkImmediately(final int chunkX, final int chunkZ) {
+        return null;
+    }
 
     public VariableBlockLightHandler getCustomLightHandler();
