+++ src/main/resources/starlight.mixins.json
@@ -13,7 +13,6 @@
     "common.chunk.EmptyChunkMixin",
     "common.chunk.WorldChunkMixin",
     "common.compat.create.WrappedChunkMixin",
-    "common.compat.create.WrappedWorldMixin",
     "common.lightengine.ServerWorldLightManagerMixin",
     "common.lightengine.WorldLightManagerMixin",
     "common.world.ChunkSerializerMixin",
